message id to 16 bit

+++ Include/CAN/CAN_Library.hpp
@@ -5,7 +5,7 @@
 #include <vector>
 
 struct Message {
-	uint8_t id;
+	uint16_t id;
 	uint16_t packet_size;
 	std::vector<uint8_t> data_field;
 };
